Update dev and sandbox backends to use updated redis IP (#347)
The redis instance in our dev environment was in a different region from our app engine instance, causing ~50ms latency for all redis lookups. Robin set up a new instance in the correct region, and this PR updates our config files to point to that instance instead of the old instance.

Roundtrip time with the new instance is ~0.5 - 1ms.