Don't delete NFT if user still owns the same NFT in a different wallet (#366)
This is a fix for an issue one of our users was running into. When doing an Opensea refresh, it's possible for the user to have multiple copies of the same NFT spread throughout their wallets. Historically, we've tried to keep the old IDs (since they're potentially used in collections), but update them with the new owner_addresses. Unfortunately, this strategy can fail when working with multiple copies of an NFT distributed among different wallets, as we might attempt to map the same new address to multiple old versions of the NFT, and that violates a uniqueness constraint in our database.

As a fix, @bennycio and @Robinnnnn and I decided not to delete the old versions of the NFT at all. If the user brings in new copies of the same NFT, we'll keep the old ones and the new ones. This potentially allows the user to have more copies of the NFT than they should, but we'll also be implementing a follow-up PR to deduplicate these on the frontend.