Always update token entries, regardless of block number (#414)
The SQL statement for bulk upserting tokens was only updating existing tokens if the incoming token had a higher block number, but that logic was actually meant for the indexer. For our purposes with updating tokens, we always want to do the update -- otherwise we'd miss cases like adding tokens with lower block numbers when connecting a wallet.